wpe: respect libs listed in egl.pc and glesv2.pc pkg-config files
By default FindEGL.cmake and FindOpenGLES2.cmake check for pkg-config
files but then ignore the list of libraries they return and always
assume libraries are called libEGL.so and libGLESv2.so.

That works OK if the the real libraries are single libraries and can
be linked against via symlinks from the expected names. However it
doesn't work if the real libraries are split into multiple separate
libraries and all of them need to be included when linking.

Warning: Experimental change, to be reviewed.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>