meta-wpe: implement initial dynamic layers support.
This fixes warnings from dangling bbappends such as `broadcom-refsw` when building for RPI machines, that do not use these recipes.

In this initial implementation, two dynamic layers are implemented: `raspberrypi` and `brcm`.
The `rpi-config`, `base-files` and `linux-raspberrypi` RPI-specific recipes/bbappends were moved (and adapted) to the `raspberrypi` dynamic layer.
The `broadcom-refsw` bbappend was moved to the `brcm` dynamic layer.
Other BSP-dependant recipes should be added to new dynamic layers accordingly.

This is work in progress towards making `meta-wpe` more BSP-independant and in turn easier to maintain overall.
More existing recipes should be adapted/moved to these dynamic layers to separate the RPI-specific parts.

Signed-off-by: Hugo Hromic <hhromic@gmail.com>