gyp: drop recipe, no longer needed.
This was original required by the webdriver component for WPE. However a lot has changed and there is no webdriver in this layer, though upstream work with webdriver is on going it is unsure what kind of dependencies the new implementation has.
Once that clarifies we'll revisit the yocto dependencies. For now, cleaning up.

Signed-off-by: wouterlucas <wouter@wouterlucas.com>