meta-wpe: normalize GitHub SRC_URI in recipes for default protocol/branch
Normalize the GitHub SRC_URI of the recipes:

    wpebackend-mesa wpebackend-rdk wpebackend wpeframework-plugins
    wpeframework-ui wpeframework wpewebkit

for accessing GitHub URIs using standard git protocol. Fixes the following warning:

    WARNING: wpeframework-ui-3.0+gitrAUTOINC+c0b3857405-r0 do_fetch:
    Failed to fetch URL git://git@github.com/WebPlatformForEmbedded/WPEFrameworkUI.git;protocol=ssh;branch=master,
    attempting MIRRORS if available

because of the fetcher trying to use SSH protocol without credentials. We can't assume
everyone has SSH keys configured, so just use the standard git protocol (the default).
Also, there is no need to specify the master branch as this is also a default in the fetcher.
Signed-off-by: Hugo Hromic <hhromic@gmail.com>