wpeframework: bump to latest and add plugin details for resume and set RDEPENDS for RPI.
This fixes the resumes: null in the main config.json, when that's not set WPEFramework will not resume any plugins.
The default assumed webserver and webkitbrowser are enabled in wpeframework-plugins. They are made available through
PACKAGECONFIG so they can be adjusted to meet other specific needs for more exotic setups.

Added RDEPENDS on userland to remove the warning that mentions a DEPENDS on userland when building WPEFramework. This is due
to certain modules having a deeper integration into RPI and is only valid for RPI targets. This is e.g. used by DeviceInfo
to retrieve GPU information from a raspberrypi.

Signed-off-by: wouterlucas <wouter@wouterlucas.com>