Cobalt: Remove unconditional DEPENDS/RDEPENDS to libcobalt
The build and runtime dependencies for libcobalt are already correctly
set by the PACKAGECONFIG for cobalt, so there's no need to have the
unconditional DEPENDS/RDEPENDS for it.

In fact, this forces the dependency even when cobalt is not activated in
PACKAGECONFIG, because cobalt.inc is always included. This is undesired
behavior as it brings unnecessary/unwanted dependencies to libcobalt
when cobalt is not activated.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>